+++ app/Crypto/Indicators/MovingAverageLatestDiffCumul.php
@@ -86,7 +86,7 @@ class MovingAverageLatestDiffCumul implements Indicator
 
     }
 
-    $cumul = (float) number_format($cumul, 2);
+    $cumul = number_format($cumul, 2);
 
     return $cumul;
